Update 20191203.1
帧率改用double类型
README更新延迟测试数据

+++ MirrorCaster/DeviceInfoData.cs
@@ -4,7 +4,7 @@
     {
         public int deviceWidth = 1920;
         public int deviceHeight = 1080;
-        public int deviceRefreshRate = 60;
+        public double deviceRefreshRate = 60;
         public bool deviceVmode = false; // 是否为竖屏状态
     }
 }
